requirement.txt
flask
login
SQLAlchemy
flask-wtf
wtforms
 langchain
 PyPDF
 python-dotenv
 streamlit
 faiss-cpu
 streamlit-extras